Create new marker based on user search box entry
Created placeMarkers variable. Add event listener to listen for the event fired when the user selects a search prediction and retrieve details for that place. Clear out the old marker on each new search and get the icon, name and location for the new place searched and create a new marker for that place